+++ blocklets/ai-studio/src/locales/en.ts
@@ -653,4 +653,7 @@ export default flat({
   previewInNewTab: 'Preview in new tab',
   copyPreviewUrl: 'Copy preview URL',
   system: 'System',
+  agentEmptyTitle: '‌No Agent Created',
+  agentEmptySubTitle: 'You can click on the "+" icon in the upper right corner to create.',
+  back: 'Back',
 });