+++ blocklets/ai-studio/src/pages/project/publish-view.tsx
@@ -1,9 +1,11 @@
-import { stringifyIdentity } from '@api/libs/aid';
 import { AI_RUNTIME_COMPONENT_DID } from '@app/libs/constants';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
+import { stringifyIdentity } from '@blocklet/ai-runtime/common/aid';
 import { AssistantYjs } from '@blocklet/ai-runtime/types';
 import ComponentInstaller from '@blocklet/ui-react/lib/ComponentInstaller';
 import { Icon } from '@iconify-icon/react';
+import CopyIcon from '@iconify-icons/tabler/copy';
+import ShareIcon from '@iconify-icons/tabler/device-desktop-share';
 import {
   Box,
   Button,
@@ -76,19 +78,15 @@ function PublishViewContent({
           },
         },
       }}>
-      <Stack gap={1} px={2} py={2}>
-        <Button
-          variant="outlined"
-          href={previewUrl}
-          target="_blank"
-          endIcon={<Icon icon="tabler:device-desktop-share" />}>
+      <Stack gap={1} px={2} py={2} className="publish-container">
+        <Button variant="outlined" href={previewUrl} target="_blank" endIcon={<Icon icon={ShareIcon} />}>
           {t('previewInNewTab')}
         </Button>
 
         <Tooltip title={copied ? t('copied') : undefined} placement="top" disableInteractive>
           <Button
             variant="outlined"
-            endIcon={<Icon icon={copied ? 'tabler:copy-check' : 'tabler:copy'} />}
+            endIcon={<Icon icon={copied ? 'copy-check' : CopyIcon} />}
             onClick={() => {
               navigator.clipboard.writeText(previewUrl);
               setCopied(true);
@@ -97,7 +95,7 @@ function PublishViewContent({
           </Button>
         </Tooltip>
 
-        <Box textAlign="center">
+        <Box textAlign="center" className="qr-code">
           <Box
             component={QRCode}
             value={previewUrl}