+++ blocklets/ai-studio/src/pages/project/test-view.tsx
@@ -1,8 +1,12 @@
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
 import { runAssistant } from '@blocklet/ai-runtime/api';
+import { stringifyIdentity } from '@blocklet/ai-runtime/common/aid';
 import { AssistantResponseType, AssistantYjs } from '@blocklet/ai-runtime/types';
 import { Map, getYjsValue } from '@blocklet/co-git/yjs';
 import { Icon } from '@iconify-icon/react';
+import BugIcon from '@iconify-icons/tabler/bug';
+import RocketIcon from '@iconify-icons/tabler/rocket';
+import TrashIcon from '@iconify-icons/tabler/trash';
 import { Error } from '@mui/icons-material';
 import { LoadingButton } from '@mui/lab';
 import { Alert, Box, Button, Stack, Tooltip, Typography } from '@mui/material';
@@ -63,7 +67,7 @@ export default function DebugView({
           sx={{ py: 0 }}
           loading={running}
           onClick={runAll}
-          startIcon={<Box component={Icon} icon="tabler:rocket" sx={{ fontSize: 16 }} />}>
+          startIcon={<Box component={Icon} icon={RocketIcon} sx={{ fontSize: 16 }} />}>
           {t('runAll')}
         </LoadingButton>
       </Stack>
@@ -135,10 +139,8 @@ const TestCaseView = forwardRef<
     try {
       const result = await runAssistant({
         url: joinURL(PREFIX, '/api/ai/call'),
-        projectId,
-        ref: gitRef,
+        aid: stringifyIdentity({ projectId, projectRef: gitRef, assistantId: assistant.id }),
         working: true,
-        assistantId: assistant.id,
         parameters: test.parameters,
         debug: true,
       });
@@ -191,7 +193,7 @@ const TestCaseView = forwardRef<
           <Tooltip title={t('runThisCase')}>
             <span>
               <Button sx={{ minWidth: 0, width: 32, height: 32 }} size="small" disabled={loading} onClick={runTest}>
-                <Box component={Icon} icon="tabler:rocket" sx={{ fontSize: 15 }} />
+                <Box component={Icon} icon={RocketIcon} sx={{ fontSize: 15 }} />
               </Button>
             </span>
           </Tooltip>
@@ -199,7 +201,7 @@ const TestCaseView = forwardRef<
           <Tooltip title={t('debugThisCase')}>
             <span>
               <Button sx={{ minWidth: 0, width: 32, height: 32 }} size="small" onClick={debugTest}>
-                <Box component={Icon} icon="tabler:bug" sx={{ fontSize: 15 }} />
+                <Box component={Icon} icon={BugIcon} sx={{ fontSize: 15 }} />
               </Button>
             </span>
           </Tooltip>
@@ -207,7 +209,7 @@ const TestCaseView = forwardRef<
           <Tooltip title={t('deleteThisCase')}>
             <span>
               <Button sx={{ minWidth: 0, width: 32, height: 32 }} size="small" onClick={deleteTest} color="warning">
-                <Box component={Icon} icon="tabler:trash" sx={{ fontSize: 15 }} />
+                <Box component={Icon} icon={TrashIcon} sx={{ fontSize: 15 }} />
               </Button>
             </span>
           </Tooltip>